+++ src/main/java/com/project/board/comment/dto/request/CommentCreateRequestDto.java
@@ -1,11 +1,9 @@
 package com.project.board.comment.dto.request;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-import java.util.UUID;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -18,6 +16,6 @@ public class CommentCreateRequestDto {
 
     private Long parentId;
 
-    private UUID memberId;
+    private Long memberId;
 
 }
