+++ src/main/java/com/project/board/comment/dto/request/CommentUpdateRequestDto.java
@@ -2,7 +2,7 @@ package com.project.board.comment.dto.request;
 
 import lombok.Getter;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 @Getter
 public class CommentUpdateRequestDto {
