+++ src/main/java/com/project/board/comment/repository/CommentRepositoryImpl.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 import java.util.List;
 import java.util.Map;
 
-import static com.project.board.entity.QComment.comment;
+import static com.project.board.comment.domain.QComment.comment;
 import static com.querydsl.core.group.GroupBy.groupBy;
 import static com.querydsl.core.group.GroupBy.list;
