+++ src/main/java/com/project/board/member/enumeration/converter/MemberStatusConverter.java
@@ -2,8 +2,8 @@ package com.project.board.member.enumeration.converter;
 
 import com.project.board.member.enumeration.MemberStatus;
 
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
+import jakarta.persistence.AttributeConverter;
+import jakarta.persistence.Converter;
 
 @Converter
 public class MemberStatusConverter implements AttributeConverter<MemberStatus, String> {
