+++ src/main/java/com/project/board/member/repository/MemberRepository.java
@@ -3,7 +3,5 @@ package com.project.board.member.repository;
 import com.project.board.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.UUID;
-
-public interface MemberRepository extends JpaRepository<Member, UUID> {
+public interface MemberRepository extends JpaRepository<Member, Long> {
 }
