+++ src/main/java/com/project/board/post/dto/request/PostCreateRequestDto.java
@@ -1,14 +1,12 @@
 package com.project.board.post.dto.request;
 
 import com.project.board.post.enumeration.PostCategory;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import java.util.UUID;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -23,10 +21,10 @@ public class PostCreateRequestDto {
     @NotNull(message = "유효하지 않은 카테고리 입니다.")
     private PostCategory postCategory;
 
-    private UUID memberId;
+    private Long memberId;
 
     @Builder
-    public PostCreateRequestDto(String title, String content, PostCategory postCategory, UUID memberId) {
+    public PostCreateRequestDto(String title, String content, PostCategory postCategory, Long memberId) {
         this.title = title;
         this.content = content;
         this.postCategory = postCategory;
